URGENT: Fix Alt < 0 wraparound in tracking (Dangerous!)

+++ indi-celestronaux/celestronaux.cpp
@@ -1409,11 +1409,11 @@ long CelestronAUX::GetALT()
     // return alt encoder adjusted to -90...90
     if (m_AltSteps > STEPS_PER_REVOLUTION / 2)
     {
-        return m_AltSteps - STEPS_PER_REVOLUTION;
+        return (long)m_AltSteps - STEPS_PER_REVOLUTION;
     }
     else
     {
-        return m_AltSteps;
+        return (long)m_AltSteps;
     }
 };
